Fixed a Null Pointer Exception Error.
This error occured when the logout button was pressed. The user would log out, but the onPause method would still be called. Because the user had logged out the userID would be null, and so giving us an error. I just made a boolean for the logged status, which fixed the problem.